+++ tests/instrumentation/urllib3_tests.py
@@ -125,3 +125,17 @@ def test_span_only_dropped(instrument, elasticapm_client, waiting_httpserver):
     assert trace_parent_1.span_id != transaction_object.id
     # second request should use transaction id as span id because there is no span
     assert trace_parent_2.span_id == transaction_object.id
+
+
+def test_url_sanitization(instrument, elasticapm_client, waiting_httpserver):
+    waiting_httpserver.serve_content("")
+    url = waiting_httpserver.url + "/hello_world"
+    url = url.replace("http://", "http://user:pass@")
+    transaction_object = elasticapm_client.begin_transaction("transaction")
+    pool = urllib3.PoolManager(timeout=0.1)
+    r = pool.request("GET", url)
+    elasticapm_client.end_transaction("MyView")
+    transactions = elasticapm_client.events[TRANSACTION]
+    span = elasticapm_client.spans_for_transaction(transactions[0])[0]
+
+    assert "pass" not in span["context"]["http"]["url"]