+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+[Check the diff](https://github.com/elastic/apm-agent-python/compare/v4.2.2...master)
+ * Added instrumentation for `urllib2` (Python 2) / `urllib.request` (Python 3) (#464)
+
 ## v4.2.2
 [Check the diff](https://github.com/elastic/apm-agent-python/compare/v4.2.1...v4.2.2)
  * fixed an issue with Celery and the prefork worker pool (#444)